Refactored wptools because it exceeded the authorized 1000 lines on a single file. Splitted into wptools and wp_theme_tools. Tests also splitted
