+++ project_types/wordpress/tests/generate_wordpress_test.py
@@ -41,9 +41,9 @@ def test_main_given_required_files_when_not_present_and_localhost_and_no_default
 @patch("logging.info")
 @patch("core.log_tools.log_indented_list")
 @patch("project_types.wordpress.wptools.get_db_admin_from_environment")
-@patch("project_types.wordpress.wptools.build_theme")
+@patch("project_types.wordpress.wp_theme_tools.build_theme")
 @patch("project_types.wordpress.wptools.install_plugins_from_configuration_file")
-@patch("project_types.wordpress.wptools.install_themes_from_configuration_file")
+@patch("project_types.wordpress.wp_theme_tools.install_themes_from_configuration_file")
 @patch("project_types.wordpress.wptools.install_wordpress_site")
 @patch("project_types.wordpress.wptools.set_wordpress_config_from_configuration_file")
 @patch("project_types.wordpress.wptools.download_wordpress")
@@ -52,7 +52,7 @@ def test_main_given_required_files_when_not_present_and_localhost_and_no_default
 @patch("project_types.wordpress.wptools.get_wordpress_path_from_root_path")
 @patch("project_types.wordpress.wptools.get_site_configuration_from_environment")
 @patch("project_types.wordpress.wptools.get_required_file_paths")
-@patch("project_types.wordpress.wptools.get_themes_path_from_root_path")
+@patch("project_types.wordpress.wp_theme_tools.get_themes_path_from_root_path")
 @patch("filesystem.paths.files_exist_filtered")
 @patch("project_types.wordpress.wptools.get_constants")
 def test_main_given_required_files_when_present_then_calls_wptools_get_required_file_paths(
@@ -82,9 +82,9 @@ def test_main_given_required_files_when_present_then_calls_wptools_get_required_
 @patch("logging.info")
 @patch("core.log_tools.log_indented_list")
 @patch("project_types.wordpress.wptools.get_db_admin_from_environment")
-@patch("project_types.wordpress.wptools.build_theme")
+@patch("project_types.wordpress.wp_theme_tools.build_theme")
 @patch("project_types.wordpress.wptools.install_plugins_from_configuration_file")
-@patch("project_types.wordpress.wptools.install_themes_from_configuration_file")
+@patch("project_types.wordpress.wp_theme_tools.install_themes_from_configuration_file")
 @patch("project_types.wordpress.wptools.setup_database")
 @patch("project_types.wordpress.wptools.install_wordpress_site")
 @patch("project_types.wordpress.wptools.set_wordpress_config_from_configuration_file")
@@ -94,7 +94,7 @@ def test_main_given_required_files_when_present_then_calls_wptools_get_required_
 @patch("project_types.wordpress.wptools.get_wordpress_path_from_root_path")
 @patch("project_types.wordpress.wptools.get_site_configuration_from_environment")
 @patch("project_types.wordpress.wptools.get_required_file_paths")
-@patch("project_types.wordpress.wptools.get_themes_path_from_root_path")
+@patch("project_types.wordpress.wp_theme_tools.get_themes_path_from_root_path")
 @patch("filesystem.paths.files_exist_filtered")
 @patch("project_types.wordpress.wptools.get_constants")
 def test_main_given_required_files_when_present_and_create_db_then_calls_setup_database(
@@ -128,9 +128,9 @@ def test_main_given_required_files_when_present_and_create_db_then_calls_setup_d
 @patch("core.log_tools.log_indented_list")
 @patch("clint.textui.prompt.yn")
 @patch("project_types.wordpress.wptools.get_db_admin_from_environment")
-@patch("project_types.wordpress.wptools.build_theme")
+@patch("project_types.wordpress.wp_theme_tools.build_theme")
 @patch("project_types.wordpress.wptools.install_plugins_from_configuration_file")
-@patch("project_types.wordpress.wptools.install_themes_from_configuration_file")
+@patch("project_types.wordpress.wp_theme_tools.install_themes_from_configuration_file")
 @patch("project_types.wordpress.wptools.install_wordpress_site")
 @patch("project_types.wordpress.wptools.set_wordpress_config_from_configuration_file")
 @patch("project_types.wordpress.wptools.download_wordpress")
@@ -139,7 +139,7 @@ def test_main_given_required_files_when_present_and_create_db_then_calls_setup_d
 @patch("project_types.wordpress.wptools.get_wordpress_path_from_root_path")
 @patch("project_types.wordpress.wptools.get_site_configuration_from_environment")
 @patch("project_types.wordpress.wptools.get_required_file_paths")
-@patch("project_types.wordpress.wptools.get_themes_path_from_root_path")
+@patch("project_types.wordpress.wp_theme_tools.get_themes_path_from_root_path")
 @patch("filesystem.paths.files_exist_filtered")
 @patch("project_types.wordpress.wptools.get_constants")
 def test_main_given_required_files_when_not_present_and_use_defaults_then_download_required_files(
